Commenting complete

+++ OpenGLAssesment/Application.h
@@ -11,12 +11,39 @@ class Model;
 class Application
 {
 public:
+	/// Setting default values
 	Application();
 	~Application();
 
-	bool Startup(float _width, float _height, const char* _title);
+	/// <summary>
+	/// Initializing glfw, create a mainwindow and call ogl_LoadFunctions()
+	/// Create a camera 
+	/// </summary>
+	/// <returns>
+	/// returns true if everything goes to plan without any error
+	/// </returns>
+	/// <param name="_width"> float </param>
+	/// <param name="_height"> float </param>
+	/// <param name="_title"> const char* </param>
+	bool Startup(int _width, int _height, const char* _title);
+
+	/// <summary>
+	/// Calculating the deltaTime and running a while loop to update every frame
+	/// </summary>
+	/// <returns>
+	/// return a bool, true if the user doesnt exit the application and false if the user
+	/// exits the application
+	/// </returns>
 	bool Run();
+
+	/// <summary>
+	/// Draws on the screen
+	/// </summary>
 	void Draw();
+
+	/// <summary>
+	/// Exits the application
+	/// </summary>
 	void ExitApplication();
 
 protected:
